refactor(web): share account sync-status derivation between sidebar headers
AccountSectionHeader and CalendarListHeader (the one-account special
case shown before any account section exists) both called
useConnectGoogle + getSidebarSyncStatus and built the same actionLabel
ternary independently. Extracted into useAccountHeaderStatus(connection),
colocated with the two components. Their JSX stays separate — one is a
collapsible per-account heading, the other a first-run CTA with a very
different button treatment — only the duplicated derivation logic moved.

No behavior change. Left getGoogleSyncStatus/getSidebarSyncStatus in
useConnectGoogle.util.ts alone: on closer inspection their overlap is
mostly shared conditions producing genuinely different copy by design
(Settings shows timestamped detail, sidebar stays terse), not the clean
duplicated-block merge the original survey suggested.
